executed "python-modernize -w -n *"

+++ crypt/rsa_key.py
@@ -31,6 +31,7 @@
 
 #------------------------------------------------------------------------------
 
+from __future__ import absolute_import
 import gc
 
 #------------------------------------------------------------------------------
@@ -117,7 +118,7 @@ class RSAKey(object):
     def verify(self, signature, message):
         h = SHA1.new(message)
         try:
-            signature_int = long(signature)
+            signature_int = int(signature)
             signature_bytes = number.long_to_bytes(signature_int)
             pkcs1_15.new(self.keyObject).verify(h, signature_bytes)
             result = True
